Make switch to MapKit and begin migration to AFNetworking 2.0
The project is transitioning through an enormous migration to AFNetworking 2.0 that has completely broken parts of the app. For instance, choosing a stop to view arrival times will instantly crash the app. Since the Live Map using MapKit instead of the Google Maps SDK for iOS is working, however, the project is being committed in this state in order to promote piecewise commits instead of one giant one when everything is working again.